Complete Refactor
